Move getCustomerRecord to a utility fn
